pyproject.toml
[project]
name = "chorez"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
scripts = { chorez = "chorez.cli.main:main" }
dependencies = [
  "dateparser>=1.2.2",
  "pydantic-settings>=2.11.0",
  "pyyaml>=6.0.3",
  "sqlalchemy>=2.0.44",
  "typed-argument-parser>=1.11.0",
]

[dependency-groups]
dev = [
  "pytest>=8.4.2",
]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["src"]
include = ["chorez*"]
